# mindforge-be/src/main/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Mindforge Demo API
  version: 1.0.0
  description: Spec-first API für Benutzer-Management und Authentifizierung

servers:
  - url: "http://{API_HOST}:{API_PORT}"
    description: "API Server"
    variables:
      API_HOST:
        default: "localhost"
      API_PORT:
        default: "8080"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegister:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    UserLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    AuthToken:
      type: object
      required:
        - username
        - token
      properties:
        username:
          type: string
        token:
          type: string

    UserInfo:
      type: object
      required:
        - id
        - username
        - role
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string

security:
  - BearerAuth: []

paths:
  /auth/register:
    post:
      summary: Benutzer registrieren
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: Benutzer erfolgreich erstellt
        '400':
          description: Ungültige Anfrage

  /auth/login:
    post:
      summary: Benutzer einloggen und JWT erhalten
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: JWT Token zurückgegeben
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '401':
          description: Ungültige Anmeldedaten

  /auth/me:
    get:
      summary: Aktuellen Benutzer abrufen
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Benutzerinformationen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Nicht autorisiert